removed some classes and constants into their own files
