+++ src/main/python/Constants.py
@@ -1,0 +1,4 @@
+INCH_TO_METER = 0.0254
+NUM_GONDOLA = 5
+NUM_SHELF = 6
+NUM_PLATE = 12
