+++ src/main/python/Coordinates.py
@@ -1,0 +1,16 @@
+class Coordinates:
+    def __init__(self, x, y, z):
+        self.x = x
+        self.y = y
+        self.z = z
+
+    def translateBy(self, delta_x, delta_y, delta_z):
+        self.x += delta_x
+        self.y += delta_y
+        self.z += delta_z
+
+    def __repr__(self):
+        return str(self)
+
+    def __str__(self):
+        return "Coordinates(%f, %f, %f)" % (self.x, self.y, self.z)
