+++ src/main/python/PickupEvents.py
@@ -2,7 +2,7 @@ from datetime import datetime
 
 import numpy as np
 
-from BookKeeper import Position
+from Position import Position
 
 
 class PickUpEvent:
